requirements.txt
requests
psycopg2-binary
apache-airflow-providers-jdbc
mysql-connector-python
boto3
pandas
